GafferArnoldUI : Move Arnold -> OSL network algo to private header
We need to use the Arnold to OSL network conversion in various
visualisers. As most are totally private and have no headers, moving
to a private helper seemed easier than allowing other visualisers
to befriend ArnoldLightVisualiser.
