+++ src/GafferArnoldUI/ArnoldLightVisualiser.cpp
@@ -34,15 +34,14 @@
 //
 //////////////////////////////////////////////////////////////////////////
 
+#include "GafferArnoldUI/Private/VisualiserAlgo.h"
+
 #include "GafferSceneUI/StandardLightVisualiser.h"
 
 #include "Gaffer/Metadata.h"
 #include "Gaffer/Private/IECorePreview/LRUCache.h"
 
 #include "GafferOSL/ShadingEngineAlgo.h"
-
-#include "IECoreScene/Shader.h"
-#include "IECoreScene/ShaderNetworkAlgo.h"
 
 #include "IECoreGL/PointsPrimitive.h"
 
@@ -51,17 +50,13 @@
 #include "IECore/TypedData.h"
 #include "IECore/VectorTypedData.h"
 
-#include "OSL/oslquery.h"
-
 #include "ai.h"
-
-#include "boost/algorithm/string/predicate.hpp"
 
 using namespace Imath;
 using namespace IECore;
 using namespace IECoreGLPreview;
 using namespace IECoreScene;
-using namespace OSL;
+using namespace GafferArnoldUI::Private;
 
 // The ArnoldLightVisualiser provides an implementation of surfaceTexture,
 // rendering a lights color input network via OSL.
@@ -73,218 +68,6 @@ using namespace OSL;
 
 namespace
 {
-
-//////////////////////////////////////////////////////////////////////////
-// OSL query LRU cache
-//////////////////////////////////////////////////////////////////////////
-
-const char *g_oslSearchPaths = getenv( "OSL_SHADER_PATHS" );
-
-typedef std::shared_ptr<OSLQuery> OSLQueryPtr;
-
-OSLQueryPtr oslQueryGetter( const std::string &shaderName, size_t &cost )
-{
-	cost = 1;
-
-	OSLQueryPtr result( new OSLQuery() );
-	if( result->open( shaderName, g_oslSearchPaths ? g_oslSearchPaths : "" ) )
-	{
-		return result;
-	}
-
-	return nullptr;
-}
-
-typedef IECorePreview::LRUCache<std::string, OSLQueryPtr, IECorePreview::LRUCachePolicy::Parallel> OSLQueryCache;
-OSLQueryCache g_oslQueryCache( oslQueryGetter, 128 );
-
-//////////////////////////////////////////////////////////////////////////
-// Network conversion helpers
-//////////////////////////////////////////////////////////////////////////
-
-// Re-connects any connections from oldName on oldShader to newName on newShader
-void remapOutputConnections( const InternedString &shader, const InternedString &oldName, const InternedString &newName, ShaderNetwork *network )
-{
-	ShaderNetwork::Parameter newSource( shader, newName );
-
-	ShaderNetwork::ConnectionRange inputConnections = network->outputConnections( shader );
-	for( ShaderNetwork::ConnectionIterator it = inputConnections.begin(); it != inputConnections.end(); )
-	{
-		// Copy and increment now so we still have a valid iterator
-		// if we remove the connection.
-		const ShaderNetwork::Connection connection = *it++;
-
-		if( connection.source.name == oldName )
-		{
-			ShaderNetwork::Parameter dest( connection.destination.shader, connection.destination.name );
-			network->removeConnection( connection );
-			network->addConnection( ShaderNetwork::Connection( newSource, dest ) );
-		}
-	}
-
-	ShaderNetwork::Parameter out = network->getOutput();
-	if( out.shader == shader && out.name == oldName )
-	{
-		network->setOutput( newSource );
-	}
-}
-
-void remapInputConnections( const InternedString &shader, const InternedString &oldName, const InternedString &newName, ShaderNetwork *network )
-{
-	ShaderNetwork::Parameter newDestination( shader, newName );
-
-	ShaderNetwork::ConnectionRange outputConnections = network->inputConnections( shader );
-	for( ShaderNetwork::ConnectionIterator it = outputConnections.begin(); it != outputConnections.end(); )
-	{
-		// Copy and increment now so we still have a valid iterator
-		// if we remove the connection.
-		const ShaderNetwork::Connection connection = *it++;
-
-		if( connection.destination.name == oldName )
-		{
-			ShaderNetwork::Parameter source( connection.source.shader, connection.source.name );
-			network->removeConnection( connection );
-			network->addConnection( ShaderNetwork::Connection( source, newDestination ) );
-		}
-	}
-}
-
-
-
-// Sets outName in outParams if inName is set in inParams.
-// Arnold data types not representable in OSL will be converted accordingly
-void copyAndConvertIfSet( const InternedString &inName,  const CompoundDataMap &inParams, const InternedString &outName, CompoundDataMap &outParams )
-{
-	const auto &it = inParams.find( inName );
-	if( it != inParams.end() )
-	{
-		if( const BoolData *boolData = dynamic_cast<const BoolData *>( it->second.get() ) )
-		{
-			outParams[ outName ] = new IntData( 1 ? boolData->readable() : 0 );
-		}
-		else
-		{
-			outParams[ outName ] = it->second;
-		}
-	};
-}
-
-// Attempts to substitute the Arnold shader with the supplied handle with an
-// OSL stand-in shader. Returns true upon success.
-//
-// If a stand-in is found, any parameters that exist on the stand-in shader
-// will be populated by the value of the equivalent parameter on the source shader.
-// Input/output connections will be remapped if required.
-//
-// The network is left un-touched if no stand-in is available.
-//
-// Stand-in shaders:
-//  - Should be placed be named: __viewer/__arnold_<shaderName>.
-//  - Should have identical parameter names/types.
-//  - Should have a single output called 'out'.
-//  - Arnold bool params should be OSL ints.
-//  - Any param name collisions with OSL reserved words should be suffixed
-//    with an '_'.
-bool substituteWithOSL( const IECore::InternedString &handle, ShaderNetwork *network )
-{
-	const Shader *arnoldShader = network->getShader( handle );
-
-	const std::string oslShaderName = "__viewer/__arnold_" + arnoldShader->getName();
-	const OSLQueryPtr query = g_oslQueryCache.get( oslShaderName );
-	if( !query )
-	{
-		return false;
-	}
-
-	ShaderPtr oslShader = new IECoreScene::Shader();
-	oslShader->setType( "osl:shader" );
-	oslShader->setName( oslShaderName );
-
-	const IECore::CompoundDataMap &aiParams = arnoldShader->parameters();
-	IECore::CompoundDataMap &oslParams = oslShader->parameters();
-	for( size_t i = 0; i < query->nparams(); ++i )
-	{
-		const OSLQuery::Parameter *parameter = query->getparam( i );
-
-		if( parameter->isoutput )
-		{
-			continue;
-		}
-
-		const std::string oslParamName = parameter->name.string();
-
-		// Skip struct members
-		if( oslParamName.find( "." ) != std::string::npos )
-		{
-			continue;
-		}
-
-		// We have to avoid collisions with function names and other language
-		// keywords, so some stand-in shaders prefix param names with '_'.
-		//   eg: normalize -> normalize_
-		std::string aiParamName = oslParamName;
-		if( aiParamName.back() == '_' )
-		{
-			aiParamName.pop_back();
-			remapInputConnections( handle, aiParamName, oslParamName, network );
-		}
-
-		copyAndConvertIfSet( aiParamName, aiParams, oslParamName, oslParams );
-	}
-
-	network->setShader( handle, std::move( oslShader ) );
-	remapOutputConnections( handle, "", "out", network );
-
-	return true;
-}
-
-// Attempts to conform the supplied shaderNetwork such that it only contains
-// OSL shaders. Arnold shaders are re-mapped in-place to their OSL equivalents
-// (if they exist). If any un-converted Arnold shaders remain, it attempts to
-// build a simple network using the first image found. If no image was found
-// then the resulting mixed network is returned.
-IECoreScene::ShaderNetworkPtr conformToOSLNetwork( const IECoreScene::ShaderNetwork::Parameter &output, const IECoreScene::ShaderNetwork *shaderNetwork )
-{
-	ShaderNetworkPtr oslNetwork = shaderNetwork->copy();
-
-	oslNetwork->setOutput( output );
-	ShaderNetworkAlgo::removeUnusedShaders( oslNetwork.get() );
-
-	bool hasUnconvertedArnoldShaders = false;
-	InternedString fallbackImageHandle;
-
-	for( const auto &s : oslNetwork->shaders() )
-	{
-		const Shader *shader = s.second.get();
-		if( boost::starts_with( shader->getType(), "ai:" ) )
-		{
-			if( !substituteWithOSL( s.first, oslNetwork.get() ) )
-			{
-				hasUnconvertedArnoldShaders = true;
-				continue;
-			}
-
-			if( shader->getName() == "image" )
-			{
-				fallbackImageHandle = s.first;
-			}
-		}
-	}
-
-	if( hasUnconvertedArnoldShaders && fallbackImageHandle != "" )
-	{
-		/// TODO: Tell the user which location this is
-		msg( Msg::Warning, "ArnoldLightVisualiser", "Unsupported shaders in network, falling back on " + fallbackImageHandle.string() );
-
-		ShaderNetworkPtr minimalNetwork = new ShaderNetwork();
-		minimalNetwork->addShader( "image", oslNetwork->getShader( fallbackImageHandle ) );
-		minimalNetwork->setOutput( ShaderNetwork::Parameter( "image", "out" ) );
-
-		oslNetwork = minimalNetwork;
-	}
-
-	return oslNetwork;
-}
 
 //////////////////////////////////////////////////////////////////////////
 // Surface texture LRU cache
@@ -322,7 +105,7 @@ CompoundDataPtr surfaceTextureGetter( const SurfaceTextureCacheGetterKey &key, s
 {
 	cost = key.resolution.x * key.resolution.y * 3 * 4; // 3 x 32bit float channels;
 
-	ShaderNetworkPtr textureNetwork = conformToOSLNetwork( key.output, key.shaderNetwork );
+	ShaderNetworkPtr textureNetwork = VisualiserAlgo::conformToOSLNetwork( key.output, key.shaderNetwork );
 	return GafferOSL::ShadingEngineAlgo::shadeUVTexture( textureNetwork.get(), key.resolution );
 }
